Resolve namespacing in a hopefully easier way
