+++ test/controllers/observations_controller/observations_controller_create_test.rb
@@ -104,7 +104,7 @@ class ObservationsControllerCreateTest < FunctionalTestCase
     assert_input_value(:herbarium_record_accession_number, "")
     assert_true(@response.body.include?("Albion, Mendocino Co., California"))
     assert_link_in_html(:create_observation_inat_import_link.l,
-                        new_observations_inat_import_path)
+                        new_inat_import_path)
 
     users(:rolf).update(location_format: "scientific")
     get(:new)
